+++ noodles-bam/examples/bam_filter_data.rs
@@ -9,13 +9,12 @@ use noodles_sam::{self as sam, record::data::field::Tag};
 
 fn is_unique_record(record: &bam::Record) -> io::Result<bool> {
     match record.data().get(Tag::AlignmentHitCount) {
-        Some(Ok(field)) => field.value().as_int().map(|hits| hits == 1).ok_or_else(|| {
+        Some(field) => field.value().as_int().map(|hits| hits == 1).ok_or_else(|| {
             io::Error::new(
                 io::ErrorKind::InvalidData,
                 format!("expected integer, got {:?}", field.value()),
             )
         }),
-        Some(Err(e)) => Err(e),
         None => Ok(false),
     }
 }